TT | Introduced `data-states` and `data-events` for multiple state and multiple event bindings on a single element in the view hierarchy. It works exactly the same and single state and event binding except that it receives an object that maps props as keys to compositions as identifiers for state and props as keys to event props as objects for events.
This allowed me to remove the need for the `DPCategories` component since I could now compose the `height` and `width` of the SVG without needing to render it inside a component function.
This allowed me to consolidate the `DOMContentLoaded` and `scroll` event bindings onto the single root element in Netflux.